+++ src/components/panel/galaxy/my/CardItem.tsx
@@ -1,4 +1,4 @@
-import Card from '@/components/common/Card';
+import Card from '@/components/common/Card/Card';
 import { IoPlanetOutline } from 'react-icons/io5';
 import { FaRegHeart } from 'react-icons/fa';
 import type { GalaxyMyListData } from '@/types/galaxyMy';
