+++ src/components/panel/stellar/objects/StellarCard.tsx
@@ -1,4 +1,4 @@
-import Card from '@/components/common/Card';
+import Card from '@/components/common/Card/Card';
 
 interface StellarCardProps {
   data: {
